split typescript projects
