+++ src/runtime/internal/animations.ts
@@ -1,4 +1,5 @@
-import { identity as linear, noop, now } from './utils';
+import { identity as linear, noop } from './utils';
+import { now } from './environment';
 import { loop } from './loop';
 import { create_rule, delete_rule } from './style_manager';
 import { AnimationConfig } from '../animate';
