+++ src/runtime/internal/index.ts
@@ -1,6 +1,7 @@
 export * from './animations';
 export * from './await-block';
 export * from './dom';
+export * from './environment';
 export * from './keyed-each';
 export * from './lifecycle';
 export * from './loop';
