+++ src/runtime/internal/transitions.ts
@@ -1,4 +1,5 @@
-import { identity as linear, is_function, noop, now, run_all } from './utils';
+import { identity as linear, is_function, noop, run_all } from './utils';
+import { now } from "./environment";
 import { loop } from './loop';
 import { create_rule, delete_rule } from './style_manager';
 import { custom_event } from './dom';
